side_project/write_s3_jsonfile_into_postgres_routine/requirements.txt
psycopg2
boto3
pytz
python-dateutil<2.8.1
